producers_orchestrator/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
idna==2.8
influxdb==5.2.2
iso-639==0.4.5
iso3166==1.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.1
MarkupSafe==1.1.1
pycryptodome==3.8.1
pydub==0.23.1
PySocks==1.7.0
python-dateutil==2.8.0
python-twitch-client==0.6.0
pytz==2019.1
requests==2.22.0
six==1.12.0
streamlink==1.1.1
urllib3==1.25.2
websocket-client==0.56.0
Werkzeug==0.15.4
